driver/usbloader: support synchronizing extra files used by uuu scripts
When using a non-builtin ``uuu`` script as image for the ``UUUDriver``,
that script can reference files relative to script file itself. To
support that case, the ``UUUDriver`` is expandend to take a list of
files to synchronize to the remote exporter before invoking ``uuu``.

However, the ``ManagedFile`` helper places the files in a per-file
directory based on the hash of the file being synchronized (when using
``rsync`` instead of ``nfs``). This breaks the files-relative-to-uuu-
script use case, so use a combined hash of all the configured files
instead.